requirements.txt
python-dotenv
ultralytics
torch
